Final Project Chapter Complete
Finished the final project chapter. It gave a good overview for how you
can create a simple async job queue system for a threaded application.
One could even probably have this at a global level and start it up
near the beginning of a program to allow for some async work outside of
what one might expect for a threaded application. Might play around
with the idea a little bit while also giving Rayon a try for some
numerical applications.